Make samples simpler (#361)
* Add dependabot configs

* Simplify crop image

* Simplify crop image java sample

* Simplify image view sample

* Simplify sample extend activity

* Rename samples files

* Refactor options dialog
