+++ sample/src/main/AndroidManifest.xml
@@ -8,7 +8,7 @@
         android:label="@string/app_name"
         android:theme="@style/Theme.Sample">
         <activity
-            android:name="com.canhub.cropper.sample.SMainActivity"
+            android:name="com.canhub.cropper.sample.MainActivity"
             android:exported="true"
             android:label="@string/app_name">
             <intent-filter>
@@ -17,7 +17,7 @@
             </intent-filter>
         </activity>
 
-        <activity android:name="com.canhub.cropper.sample.SCropResultActivity" />
+        <activity android:name="com.canhub.cropper.sample.SampleResultScreen" />
         <activity android:name="com.canhub.cropper.CropImageActivity" />
         <activity
             android:name="com.canhub.cropper.sample.extend_activity.app.SExtendActivity"
